feat: Comprehensive lint compliance & type safety improvements
- Fixed lint errors across 146 files
- Improved type annotations and imports
- Enhanced code quality with Ruff/Mypy compliance
- Updated gitignore to exclude .venv directories
- Added workflow documentation

+++ engine/src/domain/modules/layout_manager/coordinate_mapper_module.py
@@ -25,8 +25,8 @@ class CameraLike(Protocol):
 class ViewportLike(Protocol):
     """Protocol for viewport-like objects (duck typing)."""
 
-    output_width: float
-    output_height: float
+    output_width: int
+    output_height: int
 
 
 @dataclass
@@ -129,4 +129,4 @@ class CoordinateMapper:
         """
         Get the world-space rectangle currently visible.
         """
-        return self.camera.get_visible_rect(self.viewport.output_width, self.viewport.output_height)
+        return self.camera.get_visible_rect(self.viewport.output_width, self.viewport.output_height)  # type: ignore[no-any-return]
